requirements.txt
faiss_cpu==1.7.4
langchain==0.0.234
psutil==5.9.5
pydantic==1.10.11
